* win32/win32.c (rb_w32_fdcopy): suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@31411 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -2370,7 +2370,7 @@ rb_w32_fdisset(int fd, fd_set *set)
 void
 rb_w32_fdcopy(rb_fdset_t *dst, const rb_fdset_t *src)
 {
-    if (dst->capa < src->fdset->fd_count) {
+    if ((UINT)dst->capa < src->fdset->fd_count) {
 	dst->capa = (src->fdset->fd_count / FD_SETSIZE + 1) * FD_SETSIZE;
 	dst->fdset = xrealloc(dst->fdset, sizeof(unsigned int) + sizeof(SOCKET) * dst->capa);
     }
